kmssink: Save last metadata at the same time as the last buffer
The render width/height and the vinfo was only saved upon renegotiation. This
fixes the problem by saving this metadata at the same time the buffer is
saved. The saved copy of this is needed for expose() and drain() virtual functions.
This fixes various assertion that happens on drain query.

(cherry picked from commit 63940feaafa31c584a99fee006cae4f7b22146ed)
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>